Separate auth tests from books tests
